+++ src/test/scala/cromwell/util/SampleWdl.scala
@@ -2224,6 +2224,7 @@ object SampleWdl {
         |  String c = b + "MB"
         |  command { ps }
         |  runtime {
+        |    docker: "ubuntu:latest"
         |    memory: c
         |  }
         |}